backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==12.0
python-dotenv==1.0.1
python-multipart==0.0.18
aiohttp==3.11.11
pydantic==2.10.3
